+++ src/controllers/commands/builtin/chatterino/Debugging.hpp
@@ -12,4 +12,6 @@ namespace chatterino::commands {
 
 QString setLoggingRules(const CommandContext &ctx);
 
+QString toggleThemeReload(const CommandContext &ctx);
+
 }  // namespace chatterino::commands